Fix help-categorized errors: pointer ops, arg casts, enum/float/mut (total: 735→634)
Phase 1: Improve pointer type detection via field_type_map in infer_type_hint
  and is_pointer_expr_inline (Member/PtrMember/Deref support)
Phase 2: Add pointer arithmetic to inline path (.offset/.offset_from)
Phase 3: Add pointer compound assignment (wrapping_add/wrapping_sub)
Phase 4: Function argument integer type cast insertion (both paths)
Phase 5: enum comparison as u32 cast, float vs int literal, mut params

+++ tests/expected_rust/Perl_CvDEPTH.rs
@@ -3,7 +3,10 @@
 pub unsafe fn Perl_CvDEPTH(sv: *const CV) -> *mut I32 {
     unsafe {
         assert!(!sv.is_null());
-        assert!(((SvTYPE(sv) == SVt_PVCV) || (SvTYPE(sv) == SVt_PVFM)));
+        assert!(
+            (((SvTYPE(sv) as u32) == (SVt_PVCV as u32))
+                || ((SvTYPE(sv) as u32) == (SVt_PVFM as u32)))
+        );
         return (&mut (*((*sv).sv_any as *mut XPVCV)).xcv_depth);
     }
 }
